add fuzzing for wrong-code execution
This involves another testing step where the generated C program is
compiled and run two ways: one through llvm-disasm and one through the
regular clang pipeline. If the output of running those programs differs,
the test fails.

This required some heavy refactoring of the fuzzer, and adding a LLVM
version parameter for the llvm-disasm pretty printer.