Move directory constraint to cabal.project
This is an awful hack, but seems to be necessary to get SAW to build on
Windows. Stack on Windows seems to be building against multiple copies
of the same version of directory, causing builds to fail with GHCi
object loading errors. The fuzzing harness requires the latest
directory, but the rest of the package does not. Cabal doesn't do
dependency resolution in a flag-sensitive way, though, so even if the
fuzzing binary is disabled, Cabal thinks the package depends on a newer
directory package. Ugh.