Keep the `External` linkage.
I'm not clear on the purpose in dropping `External` and nothing else.
This is important information and results in LLVM bitcodes that are not
accepted when dropped.  For example the C code:

```
#include <stdio.h>
int main() { fprintf(stderr,""); return 0; }
```

Would produce:

```
@stderr_master = global %struct._IO_FILE*, align 8
```

Which is invalid (global constants need an initial value).  Instead it should be:

```
@stderr_master = external global %struct._IO_FILE*, align 8
```

This was tested with LLVM-4.0.1.  It is possible other LLVM versions
fail in which case we'll need to pass the `LLVM` constraint in and
pretty print different values in these cases (But still, the external
linkage must be known so we can do the right thing for 4.x).