Report the call stack along with exceptions (#73)
Currently, when encountering a `badValueRef`, there is an ambiguity as to which
function raised the exception. This makes such information easily accessible,
and is extensible in the sense that you can add more "HasCallStack" constraints
to get more information.

There is some duplication of information with the `ctx` variable, but I
think they serve different purposes and are both useful in debugging.