Unify assertion infrastructure
Adds an additional "Assert" module which fails with a helpful message when a
user is suspected of using an unsupported compiler.

If a record has semantically meaningful but ignored additional fields, it
can cause a soundness bug in the SAW/Crucible suite. Even the assertions
on non-semantically-important records might catch such bugs, just by
notifying the user that the developers of this project haven't looked at
that compiler yet.

Also:
 - Create a MonadFail instance for Parse
 - Add assertions to Function.hs about record lengths
 - Organize some imports