Update the parsing of the DISubprogram debug info node.
Starting with the LLVM 8.x series, this bitcode record has an even
funkier parsing method than before.  Here I've tried to mirror the
bitcode parsing logic from LLVM as closely as possible.  There are
several flags present in the LLVM data structre that are not exposed
through the LLVM pretty AST, but it would be relatively easy to add
those at a later time.