Disable failure on fcmp operator lookup failure.
DO NOT USE!  This change is only provided as a workaround for a
current activity that is only focused on recovering the callgraph from
the bitcode file.  This is a minimal patch that ignores a lookup
failure and allows a particular target to be parsed as long as it is
built with LLVM <= 9.

The core issue is a bad interpretation of the bitstream that is
resulting in an attempt to lookup fcmp operation from a value of
24 (where operations are defined for values 0-15).