Clean up treatment of fast-math flags during parsing
Currently, `llvm-pretty-bc-parser` skips over fast-math flags when parsing.
However, it was doing so incorrectly when parsing `FUNC_CODE_INST_CMP2` (i.e.,
`fcmp` instructions). Per the LLVM source code
[here](https://github.com/llvm/llvm-project/blob/b0391dfc737ede147e128fe877045f61fc5e4905/llvm/lib/Bitcode/Reader/BitcodeReader.cpp#L4342-L4377),
the fast-math flags are expected to come after all of the other fields, but the
code in `llvm-pretty-bc-parser` was assuming that they would appear in the
middle. See #158.

I addresses this issue by parsing all of the other fields upfront and avoiding
parsing the last part, which correspond to fast-math flags. While I was in
town, I also made sure to add comments in every other place in the parser where
we avoid parsing fast-math flags.