make `getBitString` assumptions explicit
There is an implicit assumption in the functions `getBitString` and
`getBitStringPartial` that they will only ever consume one additional 32-bit
word.  It should never (and it currently is never) the case that they be called
with `n` > 32.

This makes this assumption explicit in the code, so that if a later person were
to try it, they would get an explicit error, rather than some silently
incorrect results.