don't consume fast math flag entry
The record entries within a `FUNC_CODE_INST_PHI` are such that there is 1 type
entry, then 2 entries per argument, and optionally a fast math entry.

The code for consuming the pairs of argument entries was incorrectly checking
whether there remained entries, and so was attempting to read the fast math
entry as part of an argument pair, when that entry is present.

Instead, it should stop whenever there are either 0 or 1 entries remaining, as
the last entry is definitely a fast math one.