disasm-test: Implement LLVM version–specific tests
This copies over much of the version querying machinery from `crux-llvm`'s test
suite so that `tasty-sugar` can pick tests depending on if you have a
sufficiently recent version of `llvm-{as,dis}`. To test that it works as
expected, I have added a belated test case for #167 that requires LLVM 12 or
later (see `poison.at-least-llvm12.ll`).

Fixes #206.