parseAtomicRMW: Don't inspect pointee type of argument on recent LLVMs
Recent versions of the `FUNC_CODE_INST_ATOMICRMW` instruction code directly
store the type corresponding to the pointer argument, which avoids the need to
pattern-match on the pointer type.  This is required to support opaque
pointers. See #177.

Older versions of the instruction (`FUNC_CODE_INST_ATOMICRMW_OLD`) do not store
this type directly, so there were have no choice but to inspect the pointee
type using `elimPtrTo`.