cmpxchg: Compare argument types using eqTypeModuloOpaquePtrs
Because `llvm-pretty` permits opaque and non-opaque pointers to coexist, it is
possible for the first argument of `cmpxchg` to be an opaque pointer and the
second argument to be a non-opaque pointer (or vice versa). We don't want to
reject such scenarios, so we compare the types of the argument using
`eqTypeModuloOpaquePtrs`, a special form of type equality that treats opaque
and non-opaque pointers as being the same. See #177.

This requires bumping the `llvm-pretty` submodule to bring in the corresponding
changes from elliottt/llvm-pretty#110.