fuzz-llvm-disasm: Remove monad-par dependency
The primary motivation for removing this dependency is because it does not
build with GHC 9.6 at the time of writing. The secondary motivation is that
using `monad-par` isn't buying us much. It's used to parallelize the fuzzer
across multiple invocations of `clang`, but (1) the default setting is to only
use a single `clang` version, and (2) not even the CI makes use of the
parallelism. In light of this, it's doubtful that anyone is actually using the
parallelism, so let's just remove it until someone asks for it.