disasm-test: add tests from C with p0 source and verify unique metadata.
The disasm-test will now check .c files by compiling with the available version
of clang and ensure that the result can be assembled and disassembled properly.
This provides additional verification over the use of existing .ll sources
because the output from clang will contain changes associated with changes in
LLVM versions whereas the .ll sources are more static.

Also adds a test to ensure that the metadata generated has unique indices.

Adds a source to demonstrate the failure of the above (will fail during testing).