Allow parsing `METADATA_DERIVED_TYPE` from Apple LLVM
Apple's fork of LLVM uses an ever-so-slightly different bitcode format than the
upstream version of LLVM. In particular, it adds an additional record to
`METADATA_DERIVED_TYPE`, which means that we must be slightly more permissive
during parsing to allow Apple LLVM bitcodes through. See the new `Note [Apple
LLVM]` for a slightly more detailed explanation. As that Note explains, this is
good enough for now, but we may want to think about a more robust solution in
the future.

I have added an `apple-llvm.bc` regression test to `disasm-test`, but it
doesn't yet pass the test suite for unrelated reasons documented in #222. I
have added a `known_bugs` entry for this test case until that issue is
resolved.

Fixes #235.