Allow llvm-dis to fail during disasm-test runs.
The llvm-dis is run in parallel to this llvm-pretty-bc-parser to inform the user
of differences between what LLVM's disassembler and this library determine from
the bitcode file.  Differences at this stage do NOT constitute a test failure,
and are simply displayed to the user as information.

The llvm-pretty-bc-parser is a permissive parser, whereas llvm-dis is relatively
strict (constrained to the current version, and using the standard LLVM
distribution rather than variations, such as Apple's modified CLANG toolkit).
Thus, there are situations where llvm-pretty-bc-parser can parse a bitcode file
but llvm-dis will fail.  This change allows the latter to happen without causing
a hard failure in the testing.