Parse constant cmp operand IDs using absolute encoding
As noted in
https://github.com/GaloisInc/llvm-pretty-bc-parser/issues/266#issuecomment-1917189766,
the operand IDs in constant expressions use an absolute encoding. This means
that parsing constant `icmp` expressions with `getConstantFwdRefAdjustedId`
(which assumes a relative encoding) is incorrect. All forms of constant
expressions besides constant `icmp` expressions use `forwardRef` (which assumes
an absolute encoding) to parse their value operands, so constant `icmp`
expressions are the odd one out.

This patch fixes the parsing of constant `icmp` expression operands by (1)
removing the `getConstantFwdRefAdjustedId` function, and (2) switching to using
`forwardRef`.

Fixes #266.