Downgrade metadata size-related errors into warnings
Previously, `llvm-pretty-bc-parser` would produce a fatal error if it
encountered a metadata record with an unexpected size. This proves to be
extremely cumbersome in practice, however, as LLVM frequently adds new fields
to metadata records, and this causes a fair bit of headaches when attempting to
support newer LLVM versions.

This patch downgrades this class of errors into warnings. It does so by:

* Introducing a new `ParseWarning` data type that captures all types of
  parser-related warnings. (For now, the only type of `ParseWarning` is
  `InvalidMetadataRecordSize`, but we may add more in the future.)

* Adding new functions to `Data.LLVM.BitCode` (all of which end with
  `*WithWarnings`) that return `ParseWarning`s alongside the parsed `Module`.
  As such, this patch does not change any of the existing `Data.LLVM.BitCode`
  API. It is up to users to decide if they want to opt into the new API that
  offers `ParseWarning`s.

See the changes in the `disasm-test` and `llvm-disasm` test suite for examples
of how to use the new API.

Fixes #248.