Resolve forward references to metadata properly in function arguments
Previously, values of type `metadata` that appeared in function arguments
(e.g., the arguments to the `llvm.dbg.declare` intrinsic) would always be
inlined, even if the original bitcode used an indirect reference to the
metadata. In more concrete terms, if the original bitcode looked like this:

```llvm
call void @llvm.dbg.assign(metadata !1, ...)

...

!1 = distinct !DIAssignID()
```

Then after roundtripping through `llvm-pretty-bc-parser`, the bitcode would
look like this:

```llvm
call void @llvm.dbg.assign(metadata !DIAssignID, ...)
```

This not only results in unnecessarily verbose bitcode, it also poses a
correctness issue. LLVM requires that all `DIAssignID` nodes be marked
`distinct`, which can only happen if the metadata node appears in the top-level
list of unnamed metadata. If a metadata node is inlined into its use site, then
the `distinct` keyword must necessarily be dropped, thereby making the
resulting LLVM code invalid!

As it turns out, `llvm-pretty-bc-parser` was already maintaining the state
needed to construct references to metadata argument values like the one above,
but it wasn't plumbing the state to the relevant parts of the code. This patch
does just that. As part of this plumbing, we need to introduce an auxiliary
`HasMdRefTable` type class (mirroring the existing `HasMdTable`, `HasParseEnv`,
and `HasValueTable` type classes), and then use this type class to consult the
table of metadata references when parsing function arguments in
`getFnValueById'`.

Addresses problem (2) in
https://github.com/GaloisInc/llvm-pretty-bc-parser/issues/260#issuecomment-1755312754,
but not problem (1).