Refactor castOpGeneric
In particular, change `castOpGeneric`'s type signature to be more like
`binOpGeneric`, which takes a `Maybe Int` argument representing optional flags
that can be passed to certain operations. (This argument currently goes unused,
but it will be used in follow-up commits.)

No functional changes are intended as a result of this patch. This is meant to
be purely an internal refactoring in preparation for subsequent commits, where
we will need to account for `CastOp`s gaining additional flags.