Use tree-diff to report direct AST differences in disasm-test.
This will allow the direct comparison of _normalized_ AST structures instead of
just a file-based diff of the Show of those structures (which is what was done up
until this change).

Doing a direct comparison allows for normalization to address elements that are
not particularly important or are equivalent, but have different representations.
For example, LLVM 21 changes the Size and Offset fields in DIBasicType,
DICompositeType, DIDerivedType, and several others to be either a direct integer
or a metadata reference; a metadata reference with a size of 8 should be equal to
a direct integer of 8, and the normalization can address this.