Use common parsing of DebugLoc.
Reduces duplications to avoid errors, and this will be useful for LLVM 21's
additions to DebugLoc.

Note that the actual parsing routine needs to be fairly polymorphic: one call
uses `Int` as the `r` parameter and the other uses `BlockLabel`, and the helpers
for those are slightly different, thus they are passed as arguments.