Use i1, not i0, in special case for ValBool parsing
The boolean literals `false` and `true` have the type `i1`, but the special
case for `ValBool`s in the logic for parsing integer constants incorrectly used
the type `i0`. As such, `true` would never be parsed as `ValBool True`, and in
certain cases, it might even be parsed as `ValInteger (-1)`, as `-1` is one
possible encoding of `true` at the bitcode level. This, in turn, would be
pretty-printed as `i1 18446744073709551615`, which is very counter-intuitive to
users.

Fix this by special-casing `i1` instead.
Fixes https://github.com/GaloisInc/llvm-pretty/issues/13.